+++ src/main/kotlin/com/github/naoyukik/intellijplugingithubnotifications/domain/model/GitHubNotification.kt
@@ -11,12 +11,13 @@ data class GitHubNotification(
     val reason: String,
     @SerialName("updated_at")
     val updatedAt: String,
+    val detail: NotificationDetail? = null,
 ) {
     @Serializable
     data class Subject(
         val title: String,
         val url: String,
-        val type: String,
+        val type: SubjectType,
     )
 
     @Serializable
@@ -26,4 +27,29 @@ data class GitHubNotification(
         @SerialName("html_url")
         val htmlUrl: String,
     )
+
+    @Serializable
+    enum class SubjectType {
+        Issue,
+        PullRequest,
+        Release,
+        UNKNOWN,
+        ;
+
+        fun isRelevantType(): Boolean {
+            return when (this) {
+                Issue, PullRequest, Release -> true
+                UNKNOWN -> false
+            }
+        }
+
+        fun setApiPath(): String? {
+            return when (this) {
+                Issue -> "issues"
+                PullRequest -> "pulls"
+                Release -> "releases"
+                UNKNOWN -> null
+            }
+        }
+    }
 }